+++ packages/common/mdx-vite/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-- [remark-plugins]: add support for frontmatter descriptions ([c482297](https://github.com/qualcomm/qualcomm-ui/commit/c4822975506ebce79e39fe31b1c1ed3f793649c7))
+- [remark-plugins]: add support for frontmatter descriptions
 
 ### Miscellaneous Chores