+++ packages/common/mdx-vite/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @qualcomm-ui/mdx-vite
 
+## 2.8.0 (2025/12/18)
+
+### Features
+
+- [remark-plugins]: add support for frontmatter descriptions
+
+### Miscellaneous Chores
+
+- **deps:** update dependencies [@qualcomm-ui/mdx-common]
+
 ## 2.7.0 (2025/12/15)
 
 ### Features